Fixed station coordinates from int to double

+++ src/main/java/com/aes/dashboard/backend/model/Station.java
@@ -17,10 +17,10 @@ public class Station {
     private String type;
 
     @Column(nullable = false)
-    private int latitude;
+    private double latitude;
 
     @Column(nullable = false)
-    private int longitude;
+    private double longitude;
 
     @OneToMany(fetch = FetchType.LAZY, mappedBy = "station")
     private List<Observation> observationList;
@@ -49,19 +49,19 @@ public class Station {
         this.type = type;
     }
 
-    public int getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public int getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
